+++ extension/js/SaveImageDialog.jsx
